backend/requirements.txt
asgiref==3.8.1
attrs==24.2.0
aiohttp>=3.9.0
certifi==2024.8.30
channels==4.2.0
charset-normalizer==3.4.0
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
dj-rest-auth==7.0.0
Django==5.1.3
django-allauth==65.2.0
django-channels-jwt==0.0.3
django-cors-headers==4.6.0
django-environ==0.11.2
django-filter==24.3
django-rest-auth==0.9.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
environ==1.0
Faker==33.1.0
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
numpy==2.2.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
psycopg==3.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.3.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
razorpay==1.4.2
referencing==0.35.1
reportlab==4.2.5
requests==2.32.3
rpds-py==0.21.0
scikit-learn==1.5.2
scipy==1.14.1
service-identity==24.2.0
six==1.16.0
sqlparse==0.5.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
zope.interface==7.2
# langchain==0.3.11
# langchain-openai==0.2.12
# langchain-core==0.3.24
# langchain-community==0.3.11
# langgraph==0.2.59
